docs: create product workshop updates for 6-29-2022
This commit creates a blog post describing the 2nd half of our product
workshop (this post acts as the product update for that week). This
commit also adds a bunch of images used in that blog post.

Note that we should probably update the way that we're storing and
hosting blog post images. It may be nice to have them organized more
closely to the blog post in which they're used instead in the
`/public/ss` folder.

Update blog.product-workshop-second-half.mdx