chore: add tweet_processing utilities library
@nicholaschiang
I renamed the high level directory to "python" so that it makes sense to put this utility package here. lmk what you think.

This is the start of a tweet processing utility package. I am going to add an example of how to use it in upcoming commits.

I think it's best to include it in this repo, because it is going to slowly be molded by our continued experiments. It's no where near mature enough to be its own project, although that may turn into the ultimate goal