code up to moving to OO Streams
in this last update, I added a json write of a config object to manage the streams.
Now that I am doing a user login, I want to move to a Streams Object that manages all the user and recomendation, as well as the tweet fetching for the feeds and different users.